refactor backends
remove MathMethods
AlgorithmicStepMethods -> PairMethods
StorageMethods -> IndexMethods
remove most methods from StorageMethods
remove storage subpackage in backends
