+++ PySDM/backends/numba/numba.py
@@ -3,26 +3,22 @@ Created at 24.07.2019
 """
 
 from PySDM.backends.numba.impl._algorithmic_methods import AlgorithmicMethods
-from PySDM.backends.numba.impl._algorithmic_step_methods import AlgorithmicStepMethods
-from PySDM.backends.numba.impl._maths_methods import MathsMethods
+from PySDM.backends.numba.impl._pair_methods import PairMethods
 from PySDM.backends.numba.impl._physics_methods import PhysicsMethods
-from PySDM.backends.numba.impl._storage_methods import StorageMethods
+from PySDM.backends.numba.impl._index_methods import IndexMethods
 from PySDM.backends.numba.impl.condensation_methods import CondensationMethods
 from PySDM.backends.numba.random import Random as ImportedRandom
-from PySDM.backends.numba.storage.indexed_storage import IndexedStorage as ImportedIndexedStorage
-from PySDM.backends.numba.storage.storage import Storage as ImportedStorage
+from PySDM.backends.numba.storage import Storage as ImportedStorage
 
 
 class Numba(
     AlgorithmicMethods,
-    AlgorithmicStepMethods,
-    StorageMethods,
-    MathsMethods,
+    PairMethods,
+    IndexMethods,
     PhysicsMethods,
     CondensationMethods
 ):
     Storage = ImportedStorage
-    IndexedStorage = ImportedIndexedStorage
     Random = ImportedRandom
 
     default_croupier = 'local'
