--- PySDM/backends/numba/storage/__init__.py
@@ -1,3 +1,0 @@
-"""
-Created at 05.07.2020
-"""
