+++ PySDM_tests/unit_tests/particles/test_particles.py
@@ -6,6 +6,7 @@ import numpy as np
 import pytest
 
 from PySDM.storages.index import make_Index
+from PySDM.storages.indexed_storage import make_IndexedStorage
 from PySDM.state.particles_factory import ParticlesFactory
 # noinspection PyUnresolvedReferences
 from PySDM_tests.backends_fixture import backend
@@ -20,7 +21,7 @@ class TestParticles:
     def make_indexed_storage(backend, iterable, idx=None):
         index = make_Index(backend).from_ndarray(np.array(iterable))
         if idx is not None:
-            result = backend.IndexedStorage.indexed(idx, index)
+            result = make_IndexedStorage(backend).indexed(idx, index)
         else:
             result = index
         return result
